requirements.txt
attrs==18.1.0
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
django-environ==0.4.5
greenlet==0.4.14
idna==2.7
idna-ssl==1.1.0
ipython-genutils==0.2.0
jedi==0.12.1
multidict==4.3.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
yarl==1.2.6
